Change graphQl schema into modules
