+++ client/components/chapters/Sidebar.jsx
@@ -43,14 +43,12 @@ function Sidebar(props) {
           >
             <Avatar
               alt={`${user.first_name} ${user.last_name}`}
-              // src="../assets/images/broken-image.png"
               src='https://source.unsplash.com/random' // get avatar from database?
               sx={{ width: 50, height: 50, mb: 1 }}
             />
             <Typography variant="caption" sx={{ textAlign: 'center' }}>
               {`${user.first_name} ${user.last_name}`}<br />
               <Link href={`mailto:${user.email}`} underline='none'>{user.email}</Link>
-              {/* <br /> {user.phone} */}
             </Typography>
           </Box>
