+++ client/components/dashboard/GlobalDashboard.jsx
@@ -53,7 +53,7 @@ function DashboardContent({ form, chart, table }) {
     };
 
     try {
-      const response = await fetch('http://localhost:4000/graphql', options);
+      const response = await fetch('/graphql', options);
       const data = await response.json();
       if (response.ok) {
         if (mounted.current) {
