+++ client/components/map/MarkerMap.jsx
@@ -35,14 +35,12 @@ export default function MarkerMap({ lat = 40, long = -100, markerData, setSelect
       bearing: 0,
       pitch: 0,
     }}
-    // center={[long, lat]}
     latitude={coords[0]}
     longitude={coords[1]}
     onDrag={(e) => {
       setCoords([e.viewState.latitude, e.viewState.longitude]);
     }}
     style={{ width: '100%', height: '100%' }}
-    // mapStyle="mapbox://styles/mapbox/light-v10"
     mapStyle="mapbox://styles/mapbox/dark-v10"
     mapboxAccessToken={process.env.MAPBOX_ACCESS_TOKEN}
   >
