+++ server/graphqlServer.js
@@ -15,7 +15,7 @@ app.use(async (req, res, next) => {
   if (req.body.query?.split(' ')[1] === 'addChapter') {
     await chapterMiddleware.getGeocode(req, res, next);
     return next();
-  } 
+  }
   else return next();
 });
